fix: Centralise the git-mit console styling
This centralises the way we style the git error messages

Should fix #312
