+++ git-mit-config/src/cmd/lint_available.rs
@@ -1,10 +1,10 @@
 use clap::ArgMatches;
 
-use mit_commit_message_lints::lints::{Lint, Lints};
+use mit_commit_message_lints::lints::Lints;
 
 use crate::errors::GitMitConfigError;
 use crate::get_vcs;
-use comfy_table::Table;
+
 use mit_commit_message_lints::external;
 use std::env::current_dir;
 
@@ -21,24 +21,8 @@ fn run(matches: &ArgMatches) -> Result<(), GitMitConfigError> {
     let mut vcs = get_vcs(is_local, &current_dir)?;
     let toml = external::read_toml(current_dir)?;
 
-    let all_lints: Vec<Lint> = Lint::iterator().collect();
     let lints = Lints::get_from_toml_or_else_vcs(&toml, &mut vcs)?;
-    let mut table = Table::new();
-    table.set_header(vec!["Lint", "Status"]);
-
-    let rows: Table = all_lints.into_iter().fold(table, |mut table, lint| {
-        table.add_row(vec![
-            lint.name(),
-            if lints.clone().into_iter().any(|x| x == lint) {
-                "enabled"
-            } else {
-                "disabled"
-            },
-        ]);
-        table
-    });
-
-    println!("{}", rows);
+    mit_commit_message_lints::console::style::lint_table(Lints::available(), &lints);
 
     Ok(())
 }
