+++ git-mit-config/src/cmd/lint_generate.rs
@@ -22,6 +22,8 @@ fn run(matches: &ArgMatches) -> Result<(), GitMitConfigError> {
 
     let output_toml: String =
         Lints::get_from_toml_or_else_vcs(&input_toml, &mut vcs)?.try_into()?;
-    println!("{}", output_toml);
+
+    mit_commit_message_lints::console::to_be_piped(&output_toml);
+
     Ok(())
 }
