+++ git-mit/src/main.rs
@@ -7,7 +7,7 @@ use std::{
 };
 
 use clap::ArgMatches;
-use console::style;
+
 use xdg::BaseDirectories;
 
 use mit_commit_message_lints::{
@@ -17,15 +17,10 @@ use mit_commit_message_lints::{
 
 use crate::errors::GitMitError;
 use crate::errors::GitMitError::{NoAuthorInitialsProvided, NoTimeoutSet};
-use crate::ExitCode::{InitialNotMatchedToAuthor, UnparsableAuthorFile};
 use mit_commit_message_lints::mit::get_config_authors;
-use mit_commit_message_lints::mit::AuthorConfigParseError;
 
-#[repr(i32)]
-enum ExitCode {
-    InitialNotMatchedToAuthor = 3,
-    UnparsableAuthorFile,
-}
+use mit_commit_message_lints::console::exit_initial_not_matched_to_author;
+use mit_commit_message_lints::console::exit_unparsable_author;
 
 const PROBABLY_SAFE_FALLBACK_SHELL: &str = "/bin/sh";
 
@@ -41,7 +36,7 @@ fn main() -> Result<(), errors::GitMitError> {
     let config_authors = Authors::try_from(users_config.as_str());
 
     if let Err(error) = &config_authors {
-        exit_unparsable_exit_code(error);
+        exit_unparsable_author(error);
     }
 
     let all_authors = config_authors?.merge(&get_config_authors(&git_config)?);
@@ -61,27 +56,6 @@ fn main() -> Result<(), errors::GitMitError> {
     )?;
 
     Ok(())
-}
-
-fn exit_unparsable_exit_code(parse_err: &AuthorConfigParseError) {
-    let error = style("Unable to parse the author config").red().bold();
-    let tip = style(format!("You can fix this by correcting the file so it's parsable\n\nYou can see a parsable example by running:\ngit mit-config mit example\n\nHere's the technical details, that might help you track down the source of the problem\n\n{}", parse_err)).italic();
-
-    eprintln!("{}\n\n{}", error, tip);
-    std::process::exit(UnparsableAuthorFile as i32);
-}
-
-fn exit_initial_not_matched_to_author(initials_without_authors: &[&str]) {
-    let error = style(format!(
-        "Could not find the initials {}.",
-        initials_without_authors.join(", ")
-    ))
-    .red()
-    .bold();
-    let help =
-        style("You can fix this by checking the initials are in the configuration file.").italic();
-    eprintln!("{}\n\n{}", error, help);
-    std::process::exit(InitialNotMatchedToAuthor as i32);
 }
 
 fn find_initials_missing<'a>(
