+++ mit-commit-message-lints/Cargo.toml
@@ -24,6 +24,7 @@ toml = "0.5.6"
 mit-commit = "1.7.0"
 lazy_static = "1.4.0"
 console = "0.11.3"
+comfy-table = "1.0.0"
 
   [dependencies.serde]
   version = "1.0.114"
