+++ mit-commit-message-lints/src/lib.rs
@@ -2,6 +2,7 @@
 extern crate lazy_static;
 extern crate serde;
 
+pub mod console;
 pub mod external;
 pub mod lints;
 pub mod mit;
