+++ mit-commit-message-lints/src/lints/checks/body_wider_than_72_characters.rs
@@ -1,6 +1,6 @@
 use mit_commit::CommitMessage;
 
-use crate::lints::lib::Code;
+use crate::console::exit::Code;
 use crate::lints::lib::Problem;
 
 pub(crate) const CONFIG: &str = "body-wider-than-72-characters";
