+++ mit-commit-message-lints/src/lints/checks/missing_jira_issue_key.rs
@@ -1,7 +1,7 @@
 use mit_commit::CommitMessage;
 use regex::Regex;
 
-use crate::lints::lib::Code;
+use crate::console::exit::Code;
 use crate::lints::lib::Problem;
 
 pub(crate) const CONFIG: &str = "jira-issue-key-missing";
