+++ mit-commit-message-lints/src/lints/checks/not_emoji_log.rs
@@ -1,7 +1,7 @@
 use indoc::indoc;
 use mit_commit::CommitMessage;
 
-use crate::lints::lib::Code;
+use crate::console::exit::Code;
 use crate::lints::lib::Problem;
 
 pub(crate) const CONFIG: &str = "not-emoji-log";
