+++ mit-commit-message-lints/src/lints/lib/mod.rs
@@ -2,7 +2,7 @@ pub use lint::Error as LintError;
 pub use lint::Lint;
 pub use lints::Error as LintsError;
 pub use lints::Lints;
-pub use problem::{Code, Problem};
+pub use problem::Problem;
 
 mod lint;
 mod lints;
