+++ mit-commit-message-lints/src/lints/mod.rs
@@ -1,7 +1,6 @@
 pub use cmd::lint;
 pub use cmd::set_status;
 pub use cmd::SetStatusError;
-pub use lib::Code;
 pub use lib::Lint;
 pub use lib::LintError;
 pub use lib::Lints;
