+++ mit-pre-commit/src/main.rs
@@ -1,18 +1,11 @@
 use std::convert::TryFrom;
-use std::{env, process};
-
-use console::style;
+use std::env;
 
 use mit_commit_message_lints::external::Git2;
 
 use crate::cli::app;
 use crate::errors::MitPreCommitError;
 use mit_commit_message_lints::mit::get_commit_coauthor_configuration;
-
-#[repr(i32)]
-enum ExitCode {
-    StaleAuthor = 3,
-}
 
 fn main() -> Result<(), errors::MitPreCommitError> {
     app().get_matches();
@@ -23,13 +16,7 @@ fn main() -> Result<(), errors::MitPreCommitError> {
     let co_author_configuration = get_commit_coauthor_configuration(&mut git_config)?;
 
     if co_author_configuration.is_none() {
-        let error = style("The details of the mit of this commit are stale")
-            .red()
-            .bold();
-        let tip = style("Can you confirm who's currently coding?\n\nIt's nice to get and give the right credit.\n\nYou can fix this by running `git mit` then the initials of whoever is coding for example:\ngit mit bt\ngit mit bt se\n").italic();
-        eprintln!("{}\n\n{}", error, tip);
-
-        process::exit(ExitCode::StaleAuthor as i32);
+        mit_commit_message_lints::console::exit_stale_author();
     }
 
     Ok(())
